Some refactoring--only run the test if the comparison file is present, and run the test from a temporary directory.
